Fix story preview image fallback
